requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1

# PyTorch and related
torch==2.0.1
torchmetrics==1.0.3
networkx==3.1

# Configuration and logging
pyyaml==6.0.1
rich==13.4.2

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2

# Notebook support
ipython==8.14.0
jupyter==1.0.0

# Scientific computing
scikit-learn==1.3.0

# Progress bars
tqdm==4.65.0

# Graph Structure Learning
dagma==0.1.0